article-service에서 사용자 역할, 권한에 따라 동작을 수행하도록 수정

+++ article-service/src/main/java/com/kt/article/controller/ArticleController.java
@@ -3,12 +3,18 @@ package com.kt.article.controller;
 import com.kt.article.dto.ArticleRequest;
 import com.kt.article.model.Article;
 import com.kt.article.service.ArticleService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/article") // 공통 경로 설정
@@ -16,15 +22,17 @@ public class ArticleController {
     @Autowired
     private ArticleService articleService;
 
+    private static final Logger logger = LoggerFactory.getLogger(ArticleController.class);
 
 
     // 게시글 생성
     @PostMapping
-    public ResponseEntity<Article> createArticle(@RequestBody ArticleRequest request) {
+    public ResponseEntity<Article> createArticle(@AuthenticationPrincipal Jwt jwt, @RequestBody ArticleRequest request) {
         Article created = articleService.createArticle(
                 request.getTitle(),
                 request.getContent(),
-                request.getAuthor()
+                jwt.getClaimAsString("preferred_username"),
+                jwt.getSubject()
         );
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
@@ -47,15 +55,43 @@ public class ArticleController {
     @PutMapping("/{id}")
     public ResponseEntity<Article> updateArticle(
             @PathVariable String id,
+            @AuthenticationPrincipal Jwt jwt,
             @RequestBody ArticleRequest request) {
-        Article updated = articleService.updateArticle(id, request.getTitle(), request.getContent());
+        Article updated = null;
+
+        if (isAuthorization(id, jwt)) {
+            updated = articleService.updateArticle(id, request.getTitle(), request.getContent());
+        }
         return ResponseEntity.ok(updated);
     }
 
     // 게시글 삭제
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteArticle(@PathVariable String id) {
-        articleService.deleteArticle(id);
-        return ResponseEntity.noContent().build(); // 204 No Content 반환
+    public ResponseEntity<Void> deleteArticle(@PathVariable String id,
+                                              @AuthenticationPrincipal Jwt jwt) {
+        if (isAuthorization(id, jwt)) {
+            articleService.deleteArticle(id);
+        }
+        return ResponseEntity.noContent().build();
+    }
+
+    private boolean isAuthorization(String articleId, Jwt jwt) {
+        Article article = articleService.getArticle(articleId);
+        Map<String, Object> realmAccess = jwt.getClaim("realm_access");
+        List<String> roles = (List<String>) realmAccess.get("roles");
+        String currentUserId = jwt.getSubject();
+
+        if (logger.isInfoEnabled()) {
+            logger.info("[UpdateArticle] articleId : {}, authorId : {}, userId : {}, role : {}", articleId, article.getAuthorId(), currentUserId, roles.toString());
+        }
+
+        boolean isAdmin = roles.contains("kt-admin");
+        boolean isUser = roles.contains("kt-user");
+        boolean isAuthor = article.getAuthorId().equals(currentUserId);
+
+        if (isAdmin || (isUser && isAuthor)) {
+            return true;
+        }
+        throw new AccessDeniedException("수정 권한이 없습니다. (관리자 또는 작성자만 가능)");
     }
 }
